Fix the bugs in the ExpenseList and ExpenseReport screens
