--- app/src/main/java/com/assignent/expensetracker/App.kt
@@ -1,7 +1,0 @@
-package com.assignent.expensetracker
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class App : Application()
