--- app/src/main/java/com/assignent/expensetracker/data/model/Expense.kt
@@ -1,16 +1,0 @@
-package com.assignent.expensetracker.data.model
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-import java.util.*
-
-@Entity(tableName = "expenses")
-data class Expense(
-    @PrimaryKey val id: String = UUID.randomUUID().toString(),
-    val title: String,
-    val amount: Double,
-    val category: String,
-    val notes: String? = null,
-    val receiptUri: String? = null,
-    val timestamp: Long = System.currentTimeMillis()
-)
