+++ app/src/main/java/com/assignment/expensetracker/App.kt
@@ -1,0 +1,7 @@
+package com.assignment.expensetracker
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class App : Application()
