+++ app/src/main/java/com/assignment/expensetracker/data/repo/ExpenseRepository.kt
@@ -1,0 +1,11 @@
+package com.assignment.expensetracker.data.repo
+
+import com.assignment.expensetracker.data.model.Expense
+import kotlinx.coroutines.flow.Flow
+
+interface ExpenseRepository {
+    suspend fun addExpense(expense: Expense)
+    fun expensesForDateFlow(date: Long): Flow<List<Expense>>
+    fun allExpensesFlow(): Flow<List<Expense>>
+    suspend fun deleteExpense(id: String)
+}
